chore : 필요없는 파일 및 import 정리
