+++ src/components/Rotation/new_calendar.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect, useRef } from 'react';
+import React, { useEffect, useRef } from 'react';
 
 export const SelectedDates = (props) => {
   const mounted = useRef(false);
