+++ src/components/Rotation/routes.tsx
@@ -1,15 +1,13 @@
 import React from 'react';
-import { Rotate } from './Rotation';
 import { RotateResult } from './RotationResult';
 import { Route, Routes } from 'react-router';
 import RotationCalendar from './Calendar';
-
-import { NewRotate } from './new';
+import { Rotate } from './Rotation';
 
 function Rotation() {
   return (
     <Routes>
-      <Route path="/" element={<NewRotate />} />
+      <Route path="/" element={<Rotate />} />
       <Route path="/result" element={<RotateResult />} />
       <Route path="/calendar" element={<RotationCalendar />} />
     </Routes>
